feat: Add experimental strings.yml
unsure where im going with this but i thought it would be cool and convenient to have every text in one file so i dont have to go through all the code everytime to add a single thing.
for variables it uses a regex i (badly) made for a simple if else conditional "function", the rest is just normal python string.format stuff i guess.
still have stuff like Reminder Cog, Errors and a bit of contracts related stuff but otherwise most of it should be in the file